+++ webpack.config.js
@@ -31,6 +31,7 @@ module.exports = (_env, argv) => {
       vscode: 'commonjs vscode',
       ssh2: 'commonjs ssh2',
       'cpu-features': 'commonjs cpu-features',
+      'better-sqlite3': 'commonjs better-sqlite3',
     },
     plugins: [
       new webpack.DefinePlugin({ __DEV__: JSON.stringify(isDev) }),
@@ -54,6 +55,7 @@ module.exports = (_env, argv) => {
     externals: {
       ssh2: 'commonjs ssh2',
       'cpu-features': 'commonjs cpu-features',
+      'better-sqlite3': 'commonjs better-sqlite3',
     },
     plugins: [
       new webpack.DefinePlugin({ __DEV__: JSON.stringify(isDev) }),